backend/requirements.txt
# =========================
# Core Framework
# =========================
fastapi==0.110.0
uvicorn[standard]==0.27.1

# =========================
# Database & ORM
# =========================
sqlmodel==0.0.16
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# =========================
# Validation & Settings
# =========================
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.2.0

# =========================
# Authentication & Security
# =========================
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
python-multipart==0.0.9

